Intent: Strengthen trace metadata cohesion Operators involved: [Coherence, Recursivity] Affected invariants: [#8, #9] Key changes: - Centralized TypedDicts for remesh events, trace captures, and Wij payloads in tnfr.types - Updated remesh, trace, and coherence modules plus stubs to consume shared definitions - Refreshed exports to keep downstream type imports aligned Expected risks/dissonances: Minimal; typings relocated without runtime mutation Metrics: No change expected for C(t), Si, νf, phase Equivalence map: TraceMetadata, TraceSnapshot, RemeshMeta, ParallelWijPayload now sourced from tnfr.types
